+++ src/shared/index.js
@@ -1,5 +1,4 @@
 // types
-export { default as BaseProps } from './types/BaseProps';
 export { default as Color } from './types/Color';
 export { default as Size } from './types/Size';
 export { default as BaseFormProps } from './types/BaseFormProps';